Moved DataProviderSelector to com.sap.oss.phosphor.fosstars.data
Part of #588
