+++ src/test/java/com/sap/oss/phosphor/fosstars/data/DataProviderSelectorTest.java
@@ -1,4 +1,4 @@
-package com.sap.oss.phosphor.fosstars.tool.github;
+package com.sap.oss.phosphor.fosstars.data;
 
 import static org.junit.Assert.assertFalse;
